Issue#278 - Divide command package
