+++ cmd/plugin.go
@@ -1,0 +1,67 @@
+package cmd
+
+import (
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+
+	"github.com/rs/zerolog/log"
+	"github.com/spf13/cobra"
+)
+
+func createPluginCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "plugin",
+		Short: "Plugin commands",
+	}
+
+	cmd.AddCommand(&cobra.Command{
+		Use:   "status",
+		Short: "Get statuses of Plugin",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/v1/plugin_status", vatzRPC), nil)
+			if err != nil {
+				return err
+			}
+
+			cli := http.Client{}
+			resp, err := cli.Do(req)
+			if err != nil {
+				log.Error().Str("module", "plugin").Err(err)
+				return err
+			}
+
+			respData, err := io.ReadAll(resp.Body)
+			if err != nil {
+				log.Error().Str("module", "plugin").Err(err)
+				return err
+			}
+
+			statusResp := struct {
+				Status       string `json:"status"`
+				PluginStatus []struct {
+					Status     string `json:"status"`
+					PluginName string `json:"pluginName"`
+				} `json:"pluginStatus"`
+			}{}
+
+			err = json.Unmarshal(respData, &statusResp)
+			if err != nil {
+				log.Error().Str("module", "plugin").Err(err)
+				return err
+			}
+
+			fmt.Println("***** Plugin status *****")
+			for i, plugin := range statusResp.PluginStatus {
+				fmt.Printf("%d: %s [%s]\n", i+1, plugin.PluginName, plugin.Status)
+			}
+
+			return nil
+		},
+	})
+
+	cmd.PersistentFlags().StringVar(&vatzRPC, "rpc", defaultRPC, "RPC address of Vatz")
+
+	return cmd
+}
