+++ utils/version_test.go
@@ -1,4 +1,4 @@
-package main
+package utils
 
 import (
 	"testing"
